+++ src/test/java/ru/practicum/shareit/user/service/UserServiceImplTest.java
@@ -9,11 +9,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
 import ru.practicum.shareit.user.dto.UserRequestDto;
 import ru.practicum.shareit.user.dto.UserResponseDto;
+import ru.practicum.shareit.user.dto.UserUpdateDto;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repository.UserRepository;
 
-import java.util.Collections;
-import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -55,6 +54,19 @@ class UserServiceImplTest {
     }
 
     @Test
+    void updateUser() {
+        UserUpdateDto request = new UserUpdateDto();
+        request.setName("nameU");
+
+        when(userRepository.findById(Mockito.anyLong())).thenReturn(Optional.ofNullable(user));
+        when(userRepository.save(any(User.class))).thenReturn(user);
+
+        UserResponseDto response = userService.updateUser(user.getId(), request);
+
+        assertEquals("name", response.getName());
+    }
+
+    @Test
     void getUser() {
         when(userRepository.findById(Mockito.anyLong())).thenReturn(Optional.ofNullable(user));
 
@@ -63,15 +75,5 @@ class UserServiceImplTest {
         assertEquals(1L, userResponseDto.getId());
         assertEquals("name", userResponseDto.getName());
         assertEquals("w@l.com", userResponseDto.getEmail());
-    }
-
-    @Test
-    void getUsers() {
-        when(userRepository.findAll()).thenReturn(Collections.singletonList(user));
-
-        List<UserResponseDto> users = userService.getUsers();
-
-        assertFalse(users.isEmpty());
-        assertEquals("name", users.get(0).getName());
     }
 }
